fixed routing list
